+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### strukt-1.0.1.0
  - Version-control s21.sh
+ - Detect object (container) w/o explicit type
 
 #### strukt-1.0.0.3
  - Make .jar executable